requirements.txt
absl-py==0.4.0
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.14.1
Keras==1.2.0
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
numpy==1.14.5
pandas==0.23.4
protobuf==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
Theano==1.0.2
Werkzeug==0.14.1
gnsq==0.4.0
kafka-python==1.4.3
